# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test API REST",
        "description": "test",
        "version": "latest"
    },
    "paths": {
        "/api/posts": {
            "get": {
                "summary": "list all your posts",
                "operationId": "3fda362e2e082b5357ac416e7767e435",
                "responses": {
                    "201": {
                        "description": "User login successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "summary": "Register a new user",
                "operationId": "ca7c0cb6772de9bffbb8c896980cebc8",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "User’s first_name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "User’s last_name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User’s email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User’s password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "User’s phone",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "User’s image",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "login to your account",
                "operationId": "c373bc4cffa4a56fc64e3979b8ee7b96",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User’s email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User’s password",
                        "schema": {
                            "type": "password"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User login successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "logout from your account",
                "operationId": "5ff2dc8b48bff52908d3821a0ecc9b3e",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User’s email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User’s password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User login successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        }
    }
}